readme.md modified and release 0.2 completed.

+++ src/digvijag/evaluator/lib/ExpressionEvaluatorLibrary.java
@@ -26,8 +26,8 @@ public class ExpressionEvaluatorLibrary {
         return singleOperation.evaluateOperation(modifiedExpression);
     }
 
-    private String getModifiedExpression(int numberToConcat, String tempExpression, int position) {
-        String one = Integer.toString(numberToConcat);
+    private String getModifiedExpression(double numberToConcat, String tempExpression, int position) {
+        String one = Double.toString(numberToConcat);
         String two = tempExpression.substring(position, tempExpression.length());
         return one + two;
     }
